move palette file into utilities
