+++ xappt_qt/gui/utilities/dark_palette.py
